Add tzres DLL parser
